metadata backup

+++ cli/bin/csv2_metadata.py
@@ -1,12 +1,15 @@
 from csv2_common import check_keys, requests, show_active_user_groups, show_table
 from subprocess import Popen, PIPE
 
+from getpass import getuser
+from tempfile import mkdtemp
+
 import filecmp
 import json
 import os
 import yaml
 
-from csv2_group import metadata_delete, metadata_edit, metadata_list, metadata_load, metadata_update
+from csv2_group import defaults, metadata_delete, metadata_edit, metadata_list, metadata_load, metadata_update
 
 KEY_MAP = {
     '-g':   'group',
@@ -14,7 +17,7 @@ KEY_MAP = {
 
 COMMAS_TO_NL = str.maketrans(',','\n')
 
-def create_backup_file(path):
+def _create_backup_file(gvar, path, metadata):
     """
     Open the backup file (for writing), making the directory if necessary, and return the file descriptor.
     """
@@ -23,7 +26,74 @@ def create_backup_file(path):
     if not os.path.exists(path_dir):
         os.makedirs(path_dir)
 
-    return open(path, 'w')
+    if gvar['temp_dir'] and os.path.exists(path):
+        compare_path = '%s/compare%s' % (gvar['temp_dir'], path)
+        compare_dir = os.path.dirname(compare_path)
+        if not os.path.exists(compare_dir):
+            os.makedirs(compare_dir)
+
+        work_path = '%s/work%s' % (gvar['temp_dir'], path)
+        work_dir = os.path.dirname(work_path)
+        if not os.path.exists(work_dir):
+            os.makedirs(work_dir)
+        
+        p = Popen(
+            [
+                'ansible-vault',
+                'decrypt',
+                path,
+                '--output',
+                compare_path,
+                '--vault-password-file',
+                gvar['user_settings']['backup-key']
+                ],
+            stdout=PIPE, stderr=PIPE
+            )
+        stdout, stderr = p.communicate()
+
+        if p.returncode != 0:
+            raise Exception('Error: ansible-vault decrypt: stdout=%s, stderr=%s' % (stdout, stderr))
+
+        fd = open(work_path, 'w')
+        fd.write(json.dumps(metadata))
+        fd.close()
+
+        if not filecmp.cmp(work_path, compare_path, shallow=False):
+            print('Updating backup for %s' % path)
+            p = Popen(
+                [
+                    'ansible-vault',
+                    'encrypt',
+                    work_path,
+                    '--output',
+                    path,
+                    '--vault-password-file',
+                    gvar['user_settings']['backup-key']
+                    ],
+                stdout=PIPE, stderr=PIPE
+                )
+            stdout, stderr = p.communicate()
+
+            if p.returncode != 0:
+                raise Exception('Error: ansible-vault encrypt: stdout=%s, stderr=%s' % (stdout, stderr))
+
+    else:
+        fd = open(path, 'w')
+        fd.write(json.dumps(metadata))
+        fd.close()
+
+        if 'backup-key' in gvar['user_settings']:
+            p = Popen(
+                [
+                    'ansible-vault',
+                    'encrypt',
+                    path,
+                    '--vault-password-file',
+                    gvar['user_settings']['backup-key']
+                    ],
+                stdout=PIPE, stderr=PIPE
+                )
+            stdout, stderr = p.communicate()
 
 def _get_repository_and_servers(gvar):
     """
@@ -39,34 +109,70 @@ def _get_repository_and_servers(gvar):
         backup_repository = None
 
     if 'backup-repository' in gvar['user_settings']:
-        gvar['user_settings']['backup-repository'] = os.path.abspath(gvar['user_settings']['backup-repository'])
-        if not os.path.isdir(gvar['user_settings']['backup-repository']):
-            print('Error: path specified for backup repository "%s" is not a directory or does not exist.' % gvar['user_settings']['backup-repository'])
+        if gvar['user_settings']['backup-repository'] == '':
+            print('Error: path specified for backup repository cannot be null.')
             exit(1)
 
-        if gvar['user_settings']['backup-repository'] != backup_repository:
-            fd = open('%s/.csv2/backup_repository' % gvar['home_dir'], 'w')
-            fd.write(gvar['user_settings']['backup-repository'])
-            fd.close
+        else:
+            gvar['user_settings']['backup-repository'] = os.path.abspath(gvar['user_settings']['backup-repository'])
+            if not os.path.isdir(gvar['user_settings']['backup-repository']):
+                print('Error: path specified for backup repository "%s" is not a directory or does not exist.' % gvar['user_settings']['backup-repository'])
+                exit(1)
+
+            if gvar['user_settings']['backup-repository'] != backup_repository:
+                fd = open('%s/.csv2/backup_repository' % gvar['home_dir'], 'w')
+                fd.write(gvar['user_settings']['backup-repository'])
+                fd.close
+
+                os.chmod('%s/.csv2/backup_repository' % gvar['home_dir'], 0o600)
 
     elif backup_repository:
         gvar['user_settings']['backup-repository'] = backup_repository
 
+    # Retrieve default backup key - no key, no encryption.
+    if os.path.isfile('%s/.csv2/backup_key' % gvar['home_dir']):
+        fd = open('%s/.csv2/backup_key' % gvar['home_dir'])
+        backup_key = fd.read()
+        fd.close
+    else:
+        backup_key = None
+
+    if 'backup-key' in gvar['user_settings']:
+        if gvar['user_settings']['backup-key'] == '':
+            os.remove('%s/.csv2/backup_key' % gvar['home_dir'])
+            del gvar['user_settings']['backup-key']
+
+        else:
+            gvar['user_settings']['backup-key'] = os.path.abspath(gvar['user_settings']['backup-key'])
+            if not os.path.isfile(gvar['user_settings']['backup-key']):
+                print('Error: path specified for backup key "%s" is not a file or does not exist.' % gvar['user_settings']['backup-key'])
+                exit(1)
+
+            if gvar['user_settings']['backup-key'] != backup_key:
+                fd = open('%s/.csv2/backup_key' % gvar['home_dir'], 'w')
+                fd.write(gvar['user_settings']['backup-key'])
+                fd.close
+
+                os.chmod('%s/.csv2/backup_key' % gvar['home_dir'], 0o600)
+
+    elif backup_key:
+        gvar['user_settings']['backup-key'] = backup_key
+
     servers = {}
     server_xref = {}
-    for server_dir in os.listdir('%s/.csv2' % gvar['home_dir']):
-        if not os.path.isdir('%s/.csv2/%s' % (gvar['home_dir'], server_dir)):
+    for host_dir in os.listdir('%s/.csv2' % gvar['home_dir']):
+        if not os.path.isdir('%s/.csv2/%s' % (gvar['home_dir'], host_dir)) or host_dir[0] == '.':
             continue
 
-        _fd = open('%s/.csv2/%s/settings.yaml' % (gvar['home_dir'], server_dir))
-        servers[server_dir] = yaml.load(_fd)
+        _fd = open('%s/.csv2/%s/settings.yaml' % (gvar['home_dir'], host_dir))
+        servers[host_dir] = yaml.load(_fd)
         _fd.close()
 
-        server_xref[servers[server_dir]['server-address'][8:]] = server_dir
+        server_xref[servers[host_dir]['server-address'][8:]] = host_dir
 
     return servers, server_xref
 
-def _set_server(gvar, servers, server):
+def _set_host(gvar, servers, server):
     """
     Set the address and credentials for the specified cloudscheduler server.
     """
@@ -95,7 +201,63 @@ def _set_server(gvar, servers, server):
         else:
             gvar['user_settings']['server-password'] = servers['settings'][server]['server-password']
 
-    return '%s/%s' % (gvar['user_settings']['backup-repository'], servers['settings'][server]['server-address'][8:])
+    return servers['settings'][server]['server-address'][8:], '%s/%s' % (gvar['user_settings']['backup-repository'], servers['settings'][server]['server-address'][8:])
+
+def _update_git(gvar, msg):
+    """
+    Add untracked items and commit all updates.
+    """
+
+    # If the backup directory is a git repository, update git.
+    if os.path.exists('%s/.git' % gvar['user_settings']['backup-repository']):
+        p = Popen(
+            [
+                'git',
+                'status',
+                '-s'
+                ],
+            cwd=gvar['user_settings']['backup-repository'], stdout=PIPE, stderr=PIPE
+            )
+        stdout, stderr = p.communicate()
+
+        if p.returncode != 0:
+            raise Exception('Error: git status: stdout=%s, stderr=%s' % (stdout, stderr))
+
+        # Process untracked files.
+        count = 0
+        lines = stdout.decode("utf-8").split('\n')
+        for line in lines:
+            if line[:3] == ' A ' or line[:3] == ' M ':
+                count += 1
+            if line[:3] == '?? ':
+                count += 1
+                p = Popen(
+                    [
+                        'git',
+                        'add',
+                        line[3:]
+                        ],
+                    cwd=gvar['user_settings']['backup-repository'], stdout=PIPE, stderr=PIPE
+                    )
+                stdout, stderr = p.communicate()
+
+                if p.returncode != 0:
+                    raise Exception('Error: git add: stdout=%s, stderr=%s' % (stdout, stderr))
+
+        if count > 0:
+            p = Popen(
+                [
+                    'git',
+                    'commit',
+                    '-am',
+                    '%s-backup commit by %s' % (msg, getuser())
+                    ],
+                cwd=gvar['user_settings']['backup-repository'], stdout=PIPE, stderr=PIPE
+                )
+            stdout, stderr = p.communicate()
+
+            if p.returncode != 0:
+                raise Exception('Error: git commit: stdout=%s, stderr=%s' % (stdout, stderr))
 
 def backup(gvar):
     """
@@ -104,7 +266,7 @@ def backup(gvar):
 
     mandatory = []
     required = ['-br']
-    optional = ['-xA']
+    optional = ['-bk', '-xA']
     servers = {}
 
     if gvar['retrieve_options']:
@@ -120,54 +282,68 @@ def backup(gvar):
         required,
         optional)
 
-    # Retrieve data to backup For each cloudscheduler server.
+    # If the backup directory is an encrypted git repository, create a working temorary directory.
+    _update_git(gvar, 'pre')
+
+    # If the backup directory is an encrypted git repository, create a working temorary directory.
+    if 'backup-key' in gvar['user_settings'] and os.path.exists('%s/.git' % gvar['user_settings']['backup-repository']):
+        gvar['temp_dir'] = mkdtemp()
+    else:
+        gvar['temp_dir'] = None
+
+    # Retrieve data to backup for each cloudscheduler server.
+    fetched = {}
     for server in servers['settings']:
-        server_dir = _set_server(gvar, servers, server)
+        host, host_dir = _set_host(gvar, servers, server)
+        if host not in fetched:
+            fetched[host] = {}
 
         response = requests(gvar, '/settings/prepare/')
         groups = gvar['user_groups']
         for group in groups:
+            if group in fetched[host]:
+                continue
+
+            fetched[host][group] = True
+
             response = requests(gvar, '/settings/prepare/', {'group': group})
-            group_dir = '%s/groups/%s' % (server_dir, group)
+            group_dir = '%s/groups/%s' % (host_dir, group)
 
             print('Fetching: server=%s, group=%s' % (server, group))
             response = requests(gvar, '/group/defaults/')
-            fd = create_backup_file('%s/defaults' % group_dir)
-            fd.write(json.dumps(response['defaults_list']))
-            fd.close()
+            _create_backup_file(gvar, '%s/defaults' % group_dir, response['defaults_list'])
 
             response = requests(gvar, '/group/metadata-list/')
             for metadata in response['group_metadata_list']:
                 metadata_dir = '%s/metadata' % group_dir
-                fd = create_backup_file('%s/%s' % (metadata_dir, metadata['metadata_name']))
-                fd.write(json.dumps(metadata))
-                fd.close()
+                _create_backup_file(gvar, '%s/%s' % (metadata_dir, metadata['metadata_name']), metadata)
 
             response = requests(gvar, '/cloud/list/')
             for cloud in response['cloud_list']:
                 cloud_dir = '%s/clouds/%s' % (group_dir, cloud['cloud_name'])
-                fd = create_backup_file('%s/settings' % cloud_dir)
-                fd.write(json.dumps(cloud))
-                fd.close()
+                _create_backup_file(gvar, '%s/settings' % cloud_dir, cloud)
 
             response = requests(gvar, '/cloud/metadata-list/')
             for metadata in response['cloud_metadata_list']:
                 metadata_dir = '%s/metadata' % cloud_dir
-                fd = create_backup_file('%s/%s' % (metadata_dir, metadata['metadata_name']))
-                fd.write(json.dumps(metadata))
-                fd.close()
+                _create_backup_file(gvar, '%s/%s' % (metadata_dir, metadata['metadata_name']), metadata)
+
+    _update_git(gvar, 'post')
 
 def delete(gvar):
-    metadata_delete(gvar)
+    return metadata_delete(gvar)
 
 def edit(gvar):
-    metadata_edit(gvar)
+    return metadata_edit(gvar)
+
+def group(gvar):
+    return defaults(gvar)
 
 def list(gvar):
-    metadata_list(gvar)
+    return metadata_list(gvar)
 
 def load(gvar):
-    metadata_load(gvar)
+    return metadata_load(gvar)
 
 def restore(gvar):
     """
@@ -189,5 +365,5 @@ def restore(gvar):
         optional)
 
 def update(gvar):
-    metadata_update(gvar)
+    return metadata_update(gvar)
